fix(cli): Add `ELECTRIC_WRITE_TO_PG_MODE` option and fix default values (#913)
Fixes https://github.com/electric-sql/electric/pull/913

I've added `ELECTRIC_WRITE_TO_PG_MODE` as an option to the CLI to be
able to put it in `direct_writes` mode.

Changed `DATABASE_USE_IPV6` and `ELECTRIC_USE_IPV6` to be `true` by
default as per the docs.

Also changed `DATABASE_REQUIRE_SSL` to be `true` and `AUTH_MODE` to be
`secure` by default as this is the default for Electric as well, but I
do understand that we might want them to be `false` and `insecure` for
easier setup. Let me know if I should roll those back.

Finally, I've added a small test suite that really just re-confirms that
the environment variables are present and some defaults are as expected,
mostly to force us to update both the code and the test when touching
the CLI. I find this safeguard to help, but happy to remove it.