feat(client): Abstract subscription to query results away from frameworks (#884)
I've lifted the idea of subscribing to a live query to the core client
level, which simplifies writing and maintaining bindings for the various
front-end frameworks (React, Svelte, Solid, Angular, Vue). It also
allows for using this pattern outside of a framework for more
state-related business logic (e.g. for offline persistent RPC like in
the request/response pattern)

I've opted for adding to the `LiveResultContext` a `subscribe` method,
that takes a `handler` which receives either a result or an error, and
an update timestamp. I've kept the schema backwards compatible with our
existing React bindings.

Will add tests to the `createQueryResultSubscribeFunction` function but
figured I'd check if you agree with the approach first.