feat(client): Add expo-sqlite/next driver (#867)
Adds the [expo-sqlite
(next)](https://docs.expo.dev/versions/latest/sdk/sqlite-next/) driver
on top of the regular one.

I can update the docs as well but I was unsure if we wanted it as a
separate entry to the regular Expo one or if I should just add an
additional section for using `next` instead of the regular one.

Also, is there a process/tooling in place for us to do a proper test for
some of these drivers (e.g. spin up android/ios machines in CI and run
basic e2e tests) to ensure they work?