chore(electric): Make sure AUTH_MODE=secure when secure auth options are used (#949)
Fixes VAX-1589.

I can configure Electric with AUTH_MODE=insecure and AUTH_JWT_ALG=RS256
and it'll run just fine. This is dangerous because 1) I forgot to
configure the signing key, 2) it doesn't matter anyway because the token
signatures won't be validated in the insecure mode.

This change makes it so that if any secure auth settings are used together
with AUTH_MODE=insecure, Electric will fail to start and will log an
error.