feat(client): Vue bindings (#961)
Added two main components for the Vue bindings:

### `useLiveQuery`

1. Takes a static live query (e.g. `useLiveQuery(db.items.liveMany())`)
which does not update reactively, so for example
`useLiveQuery(db.items.liveMany({ where: { value: foo } })` will not
update reactively with changes to `foo`.

2. Takes a getter or reference to a live query (e.g. `useLiveQuery(() =>
db.items.liveMany())`) which allows the query itself to change
reactively with changes to its arguments. Under the hood it checks if
the provided is a reference or a getter and uses `computed` in the case
of the latter.

For both ways there is an optional seconds `runQueryDependencies`
argument which is a list of `WatchSource`s, and if provided the query
will re-run when those dependencies change (and not when the query
itself changes). This is consistent with the React bindings but I'm
unsure if this is idiomatic for Vue.

For method (1), with the static live query, it might also cause
unexpected bugs for users who expect it to update reactively so we might
want to remove that option. I kept it in for the sake of having a less
verbose option for queries that do not need that reactivity.

### `makeElectricDependencyInjector<S extends ElectricClient>`
This generates two methods, `provideElectric` and `injectElectric`, that
can be used to provide an electric instance throughout the app. Similar
to our React bindings, we delegate creating the injection to the
developer such that it can take the generated client as a type argument.


I decided to not include a `useConnectivityState` API as we are about to
break that API for the React bindings as well so we can add that later.