feat(examples): Tauri SQLite example app (#891)
This PR adds a driver for Tauri SQLite (fixes VAX-1595).
It also adds an example Tauri application.

To test the example app it needs to run the modified version of our
ts-client.
To this end, `npm pack` it and modify the `electric-sql` dependency of
the Tauri example app to point to it.

Then follow the steps in the example app's readme but follow this
deviation:
**after generating the client** go into the generated client
`src/generated/client/index.ts` and add `// @ts-nocheck` at the top of
the file. This is needed because there seem to be type errors in the
generated client (used not to be the case before, need to investigate as
a separate issue). Afterwards, you can continue with the steps from the
readme.

---------

Co-authored-by: Sam Willis <sam.willis@gmail.com>