fix(client): Minimize Web API dependencies and provide polyfills (#1009)
In the last update we introduced the `jose` dependency for JWT token
singing and decoding, which depends on the `crypto` API which is not
available in some environments (e.g. react-native).

I've tried to remove as much web API dependency or at least polyfill
them to make integration easier:

- Asynchronously import `jose` dependency for new `secureAuthToken`
method which will still fail as it requires `crypto` but we have not
documented the method yet so we can sort it out separately @kevin-dp
- Implement custom `jwtDecode` method based on the [npm package of the
same name](https://www.npmjs.com/package/jwt-decode) - reason not to add
another dependency is because it polyfills base64 with a different
base64 dependency than the one we already use, and it's fairly thin and
easy to maintain.
- Use our custom `base64` utility for encoding and decoding everywhere,
to avoid using web APIs `atob` and `btoa` - we were already using a
non-web API implementation so this is not an extra burden.
- Provide TextEncoder polyfill globally (i.e. always try to polyfill it)
with
[`fastestsmallesttextencoderdecoder`](https://www.npmjs.com/package/fastestsmallesttextencoderdecoder)
rather than adding it on the driver entry points - there's separate code
(like in the /auth utils) that uses text encoding and import order can
cause issues - we already had this as a dependency and it's fairly
light, I think it's safe to include.
- Remove `uuid` dependency and `setUUIDImpl` method and use the
`genUUID` utility everywhere - it falls back to using `Math.random` for
rng which does not give very strong uniqueness guarantees so I added a
`console.warn` to provide a polyfill.
- [BREAKING] Add
[`react-native-get-random-values`](https://www.npmjs.com/package/react-native-get-random-values)
peer dependency for expo and react-native drivers so that `genUUID`
works properly (see above point).

With these changes, react-native developers only have to include
`react-native-get-random-values` as an extra dependency (and we could
make that optional as well if we are ok with `Math.random` as our UUID
source). We can also add it as a proper dependency and only include it
in our react-native driver entry points to reduce friction even more.