fix(client): Remove comments from client SQLite trigger migrations (#1031)
We already remove the comments from the actual statements that we run on
the drivers in the [schema
migrations](https://github.com/electric-sql/electric/blob/main/clients/typescript/src/migrators/schema.ts),
so I've made the trigger installation do the same.

Not exactly an issue but there is some inconsistent comment support in
some drivers and this at least works around those issues in the initial
Electric installation. Reducing the surface of required compliance to
the SQLite spec to get Electric to run unblocks the ability to at least
try out the client with different drivers.

See [this
issue](https://github.com/capacitor-community/sqlite/issues/521) I've
opened for the Capacitor driver which motivates this PR.

Removing these comments outside of the executed statements makes our
Ionic demo with the Capacitor driver work.