fix (client): fix generic serial database adapter to always release the lock (#1033)
This PR fixes a bug with the serial database adapter that caused the
lock not to be released. This could happen when `COMMIT` failed in the
`finally` block (e.g. due to a FK constraint violation) which would
cause an error to be thrown and execution would never reach the call to
`release()`. We fixed this by moving the `COMMIT` inside the `try` block
such that any exception thrown by commit is caught. If `COMMIT` fails we
will `ROLLBACK` in the catch block and we will always release the lock
in the `finally` block.