fix (client): fix snapshot query that sets the cleartags (#1035)
This PR fixes a problem with `performSnapshot` which caused a deleted
row to re-appear in the following scenario:
TX1: `insert row`
TX2 (causally after TX1): `update row; delete row`

After TX2 the deleted row would re-appear with the value of the update.
This occurs because `performSnapshot` would only set the clearTags on
the first operation in each transaction which caused the `delete` not to
override the `update`. As discussed with @icehaunter, the solution is to
modify `performSnapshot` such that it sets the clearTags of all
operations in a TX (except the first) to the tag of that transaction.

This PR applies the above fix and i can confirm that the anomaly is
gone. I have added a unit test that checks this specific case.
**However, another unit test**, called "origin tx (INSERT) concurrently
with local txses (INSERT -> DELETE)" in `process.tags.test.ts` **started
failing**. I investigated a bit further but that unit test seems weird
to me because on line 564 it says "For this key we receive transaction
which was older" and it contains a tag that keeps the row alive (but our
own, later tx already deleted that row, see L554). So i'm not sure if it
makes sense that the server would ever send us this old transaction. I
don't know if it was meant like this or it was meant to send the ack of
our original tx in which case the tags should be empty because we
deleted that row (L572, and also delete L587).

I have modified the tags in that failing unit test according to what i
believe the server would send.