feat!: shape subscriptions (#899)
Main feature of this commit is the support for arbitrary FK traversal
when requesting shape subscription. This makes previous "multiple full
table" requests invalid, because when a table is requested without
following it's FKs, we cannot guarantee FK consistency of the client's
data view.

When requesting a shape, the server will do it's best to fulfill "at
least" the requested shape: if there are any outgoing FKs from the
tables mentioned in the requested include tree, those will also be
traversed and sent to the client.

Another important change is introduction of "additional" transactions:
those are non-timestamped transactions that the server sends to the
clients when they need to see some data they haven't seen before, but
those rows weren't affected directly. They are tied to a "real"
transaction that caused this additional data to appear by a reference
number, so the client may wait for the full view of the data if it so
chooses. Only one-to-many relations are followed with a lag like that,
many-to-one relations will always be present in the transaction itself
to ensure FK consistency on the client.

This functionality is supported by a sent rows FK graph that's
maintained per client in the WebSocket process. It may be persisted, but
should probably have a similar lifetime to WAL cache window, since
that's the current boundary for client reconnection.

Since the client can reconnect at an arbitrary point in time and we need
the same sent rows graph as the client has, this also introduces
acknowledgement messages to limit the window in which we need
to be able to restore SR graph at an arbitrary point.

---
Closes VAX-1260, VAX-593