chore(e2e_client): Don't create ts declarations in the E2E client build. (#1102)
Type declarations for the E2E client take 1.5 min to generate in the E2E
Satellite client, which turns into a slow Docker image build for the E2E
tests. Those declarations are not needed at runtime in the tests, and
`tsmodule` can be configured to skip them during build.
With these changes, the E2E Satellite client build is essentially
instant.

The reason for the slow generation might be the complexity of the client
model and zod types, but I'm not totally sure.