chore(docs): Add a disclaimer to the "Local migrations" section in the docs (#1147)
There is a lot of text between the top of the page and the "Local
migrations" section. Someone landing on the page scolled down to the
section will not see the disclaimer at the top. Moreover, support for
local migrations is not even "work in progress" as it's currently only a
planned feature.

I think this is sufficient argument to justify a separate disclaimer
specifically for this section.

Prompted by a [discussion in
Discord](https://discord.com/channels/933657521581858818/1208902100801560676/1226891278180941824).