fix(client): Don't leave a snapshot running when stopping the Satellite process (#1179)
This race condition was discovered on
https://github.com/SkillDevs/electric_dart/issues/21
A unit test is running, and it can happen that electric and the database
connection are closed, while a `performSnapshot` call is
scheduled/already running. An error is thrown when the snapshot tries to
operate with the closed database.

The easy fix would be to "wait" for the snapshot mutix after cancelling
the throttle, to ensure that it has ended to gracefully stop.

Another thing discovered is that `stop` was not returning a Promise,
just calling `_stop` under the hood.