fix(client): handle null values in BigInt columns (#1190)
This PR fixes a bug where the Satellite process is trying to parse
`null` values into a `BigInt`, as reported [on
Discord](https://discord.com/channels/933657521581858818/1229458801799729233/1229458801799729233).
I've extended a unit test to check BigInt values and also extended the
e2e test for integers to check this corner case. The e2e test would
reproduce the bug and confirms that it is solved by this fix.