fix: make sure PKs in any positions work with compensations
Original bug was that values from PKs were put in different columns.
Reason for that bug is sneaky: I used a `%ROWTYPE` variable to store
results of a `SELECT`, but `SELECT` selected PKs of the table first, and
then all other columns. I assumed that `SELECT ... INTO` would fill the
variable based on the column names, given that they match column names,
but instead it used a column order to match fields, so PK values got
into first columns of the table, and all columns up to PK were "moved"
one over.

The solution here is to use a generic `RECORD` variable instead, because
we're referring to columns by name later, and the way `SELECT` is built,
it guarantees that every column of the original table is present.