chore(starter): Test CLI more thoroughly for dep installation failure (#1207)
We had an issue with some peer deps not being marked as optional that
started causing the npm install to fail (see
https://github.com/electric-sql/electric/pull/1205).

Despite that because the starter cli does not return a non zero exit
code if it fails at the dep installation phase, some tests were passing
- I've modified them to check explicitly that the dep installation error
is not shown.

Until we upgrade the electric-sql package with the optional peer dep
flag, I've added the `--legacy-peer-deps` to `npm install` calls to work
around this issue. I will roll this back immediately after release after
confirming that the individual examples work correctly as well.