chore (client): add Zod and Prisma as (optional) peer dependencies (#1187)
Every Electric app needs Zod in order to use the generated client.
However, in recent versions of Zod something changed that breaks the
types in the generated client (cf.
https://github.com/electric-sql/electric/pull/700). Therefore, Electric
apps should always use Zod version 3.21.1. Added this as a peer
dependency such that npm complains if the app installs a different
version. Similarly, added an optional peer dependency for Prisma v4 such
that it complains if the app installs another version of Prisma.