Issue client reset message when permissions change (#1218)
Rather than attempt to reset the server connection's state after we've
asked the client to reset itself, we just close the connection. The
client will then re-connect automatically and, since it's had its state
reset, the server will treat it as a first connection (including
dropping all client reconnection state).

Since this is intended as an enabling simplification to maintain
consistency between client and server in the simplest way possible,
pending a more efficient, less brutal but much more complicated
solution, I've tried to keep the changes to the protocol as small as
possible.

Hence the use of `throw` -- it's either that or restructure all the
message handling on the server side to be more of a `reduce_while`
style.