Linearlite with PGlite (#1164)
Enables using Linearlite with PGlite or wa-sqlite by setting the
`ELECTRIC_CLIENT_DB` env var to either `wa-sqlite` or `pglite`. This can
be done in the `.env` file or any other standard way. It will be
embedded into the build so a build will be for one or the other.

To be merged after #1038 with the `electric-sql` rep either pointing at
a canary build or a published version.