fix: disable automatic fallback from SSL to non-SSL DB connection
`Postgrex` connector we're using now doesn't support fallback behavior like this, so we're disabling it for `epgsql` connections as well.
This commit also fixes application of the default value for requiring SSL, which was erroneously "false" before. This means that if your DB doesn't support SSL, you need to explicitly specify it via `DATABASE_REQUIRE_SSL=false` env variable or `?sslmode=disable` in the connection string.