fix(e2e_client): Fixes some typings in the e2e satellite client (#1222)
I've looked into why Lux didn't catch this. It appears to be because it
was matching the same string that appears async in stdout via a
SatOpLog. So I've tweaked the expectation so that it is more strict.

Another minor issue we found in the e2e client is the prisma definition
for the enums table. We generate the Dart e2e client from the schema
file, and the Enums model didn't include a `@@map`, so the generated
table name was uppercase. But the e2e test creates the enums table as
lowercase.

There appears to be other typing issues in the client, but I'm not sure
how to resolve.


![image](https://github.com/electric-sql/electric/assets/22084723/8185536b-af2b-45b4-a1cf-b32fed1b82b5)


![image](https://github.com/electric-sql/electric/assets/22084723/347f0c8f-b7bd-449f-b381-b2c0b9272a7d)

cc @kevin-dp