fix(client): Better throttle snapshot stop strategy  (#1251)
We've encountered a race condition that can occurr if, after cancelling
the throttle when stopping the process, some other piece of code calls
throttlesnapshot. That's because cancelling the throttle doesn't prevent
from scheduling new ones.

The proposed solution is to assign undefined to the function, so places
that would normally trigger a snapshot, like the poll interval, would
just do nothing until properly closed.

On top of this, we found that `setClientListeners` is only called when
instantiating the process, but it's cleaned up when calling `stop` with
the `disconnect` function. This makes it so that stopping and then
starting the process won't have some of the client listeners properly
set up. The proposed solution is to just move the client listeners
intialization to the `start` function.

I'm not sure how to test this. It would probably need some way to mock
client events after a stop and a start.