fix(client_test): Improve flakyness of shape lifecycle test  (#1288)
It looks like the new test `notifies for shape lifecycle` can fail with
the postgres driver. When expecting 1 notifications, 2 were sent
instead. This seems to be because the subscription data is emitted
before the `Process.subscribe` ends (there is an `await setMeta` that
can take a bit longer than the 1 ms delay the mock uses, when using the
Postgres driver.

I don't know if there is better way to solve this.

cc @msfstef