fix(client): Rethrow async event emitter errors ASAP (#1312)
Minor correction to the fix implemented in
https://github.com/electric-sql/electric/pull/1247

Reintroducing the `queueMicrotask` throwing approach to ensure errors
are definitely not suppressed (e.g. they need to be thrown independently
of all listener promises `settling`/fulfilling) and ideally if more than
one of them occur, all of them are exposed to the global error handler.
See discussion here
https://github.com/electric-sql/electric/pull/1247#discussion_r1620774644

The `queueMicrotask` API seems to be widely supported by browsers and
runtimes so I don't think there's a need to worry. The alternative is
also to use `setTimeout` which is also widely supported, but that might
cause the errors to be thrown significantly later than when they were
emitted.