fix(client): Fix URL parsing in the CLI (#1296)
This is just to demonstrate the direction in which URL parsing can be
improved in the CLI.

First, more test coverage is needed for some not-too-uncommon edge
cases.

Second, more uniformunity is needed in the parsing code. Currently, in
`src/cli/utils.ts` alone the database URL is parsed using a regexp, the
service URL is parsed using the deprecated `url.parse()` method. All of
this can be replaced with the built-in `URL` class and its
properties/method. That class can also be used for building URLs,
instead of ad-hoc string concatenation that doesn't escape URL
components.

---------

Co-authored-by: msfstef <msfstef@gmail.com>