feat (client): flag to disable FKs on incoming TXs in SQLite (#1281)
This PR introduces an additional flag that can be enabled when
electrifying a SQLite database. When the flag is enabled, the Satellite
process disables FK checks on incoming transactions. This flag is useful
for performance reasons as skipping FK checks greatly speeds up initial
data sync when there is a lot of data.

Note that disabling FK checks can't be done transactionally so we have
to disable FKs, apply the TX, and then re-enable FKs. Therefore, it is
important that no transactions are ran concurrently (this is ensured by
the mutex that our drivers hold).
We assume that Electric exlusively owns the DB connection and thus that
nobody executes a concurrent TX directly on the DB connection.