Check permissions for incoming operations (#1339)
As part of the move to read-only permissions this adds in permissions
checks for writes from the clients. With this in place and a later pr to
disable GRANTing write permissions, we can enable read-only permissions
without needing to do the full implementation of scoped write
permissions (which we need to figure out). If you can't grant write
perms then we don't need to resolve them.

Note for reviewers: The permissions feature flag naming ceases to make
sense with this pr, but that's all sorted in the next one, #1347