fix(cli): Use an IPV6-enabled network in CLI's compose file (#1299)
This PR changes two things:

1. Removes one of the compose files, reducing duplication. Now we can no
longer get into a situation where the sync service definition is updated
in one file but not in the other one.
2. Adds an IPv6-enabled network definition to the compose file and makes
the electric service use that network.

### Background 

Beyond requiring support in the OS, the network router and the ISP,
there are two more problems with using IPv6 for database connections:

* it's not automatically enabled in Docker
* an explicit network configuration needs to be added to the compose
file when the sync service is started by `docker compose`

The first problem is relevant to those users who utilize `docker run` to
start the sync service. The second problem currently makes our CLI
unusable with IPv6 because the compose file bundled with it has no
provisions for IPv6 networking. This PR address the latter.

### Testing

I have tested connectivity from the sync service to a Supabase database
over IPv6 and to a DigitalOcean Managed PostgreSQL over IPv4. Here's
what I've learned:

* Once IPv6 support has been enabled in the Docker daemon, it seems to
persist even after the configuration file is removed and the daemon is
restarted. For example, with IPv6 disabled in the Docker daemon (such
that `docker run ... curlimages/curl -6 https://ifconfig.co/ip` fails to
connect to the remote host) it is still possible to define an
IPv6-enabled network in a compose file and the services started by
`docker compose` will be able to connect to remote hosts over IPv6.
* The compose file may include a minimalistic network definition that
only includes the `enabled_ipv6: true` line. But for this to work, the
Docker daemon must be configured with an IPv6 default pool. Otherwise,
the network definition in the compose file has to include a subnet
definition.
* Alternatively, the network definition in the compose file may be
marked with `external: true` to instruct `docker compose` to use a
user-created Docker network instead of creating a new one.
* The CIDR prefix configured in Docker daemon's default pool(s) and the
subnet in the compose file do not need to match.
* An IPv6-enabled network also supports IPv4 connectivity.

In summary, including a default IPv6-enabled network definition in our
packaged compose file should work regardless of how users have their
Docker daemon configured.

There is a potential for conflict, though. If there's already a Docker
network on the machine that has an overlapping subnet with the one we
define in our compose file, `docker compose up` will fail. To minimize
the chances of conflicts, I have chosen a subnet with randomly generated
groups that has the size of only 4 address. To make it fully
conflict-proof, the compose file uses an environment variable that
allows for switching the host network mode or using an external,
user-created Docker network by name.

Fixes #863.